Removed generate, generateObjJ, format and ident options.
We are always going to use a format. We always want to generate from
the AST, and the code formatter will be a separate executable.

This change allowed us to dramatically simplify the code and make it much
easier to see the flow of control.

Indent options are part of the format now.