Finally got formatting right (I hope!).
Since all nodes in a format are responsible for adding linefeeds before, the only relevant information is what the previous statement at the same level is. If a node is a direct child of another statement, in the format the previous statement is now "*parent".

My apologies to Martin Carlberg! Now I understand why he was keeping a list of statements at the current level.