Added scope to identifier issues
This is necessary to ensure that the filtering of those issues occurred only within the context of the scope in which the issue occurred.

For example, an unknown identifier could be used in one function, then the same identifier could be defined in another function. Previously, the second function would filter out the issue from the first function because the issue filtering was not limited to the scope of the first function.